Fix wrong rounding with double/float values on x/y (#130)
* chore: add pre-commit and pre-push git hooks

* fix: change x and y attributes from Int to Float in SvgRectNode and related methods

* chore: bump up version

* ci: add detekt check on pull_request workflow

+++ svg-to-compose/src/commonMain/kotlin/dev/tonholo/s2c/domain/StrokeDashArray.kt
@@ -27,8 +27,8 @@ private enum class StrokeDashDrawDirection(
 }
 
 fun StrokeDashArray.createDashedPathForRect(
-    x: Int,
-    y: Int,
+    x: Float,
+    y: Float,
     width: Int,
     height: Int,
     strokeWidth: Int,
